+++ app/registries/Registry.js
@@ -1,12 +1,6 @@
 const rp = require('request-promise-native');
 const Component = require('../registry/Component');
-const { summary } = require('../prometheus');
-
-const summaryGetTags = summary({
-    name: 'wud_registry_response',
-    help: 'The Registry response time (in second)',
-    labelNames: ['type', 'name'],
-});
+const { getSummaryTags } = require('../prometheus/registry');
 
 /**
  * Docker Registry Abstract class.
@@ -72,7 +66,7 @@ class Registry extends Component {
         const getTagsOptionsWithAuth = await this.authenticate(image, getTagsOptions);
         const response = await rp(getTagsOptionsWithAuth);
         const end = new Date().getTime();
-        summaryGetTags.observe({ type: this.type, name: this.name }, (end - start) / 1000);
+        getSummaryTags().observe({ type: this.type, name: this.name }, (end - start) / 1000);
         return response;
     }